+++ backend/problems/src/ma1/algebra/expressions_without_parentheses.rs
@@ -2,7 +2,6 @@ use anyhow::Result;
 use macros::problem;
 use math::{Evaluable, Number, Polynomial, Term, VariableList, num_gen, symbols};
 use rand::seq::IndexedRandom;
-use registry::get_problem_data;
 use types::{format_strings::HasReplacements, lang::Language, problems::Problem};
 
 /// 3x + 4 + 2x + 1
@@ -176,10 +175,7 @@ fn evaluate_simple(id: i32, lang: Language) -> Result<Problem> {
         ("unknown", unknown.to_string()),
         ("value", value.to_string()),
     ];
-    let problem_data = get_problem_data(id)?;
-    let question = problem_data
-        .get_question(lang)
-        .replace_placeholders(&replacement_map);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&replacement_map);
     let replacements = [(unknown, &value)];
     let answer = expression.evaluate(&replacements);
 
@@ -221,8 +217,7 @@ fn evaluate_intermediate(id: i32, lang: Language) -> Result<Problem> {
     let const_term = Term::from_num(constant);
 
     let expression = Polynomial::from_terms(&[&first_term, &second_term, &const_term]);
-    let problem_data = get_problem_data(id)?;
-    let question = problem_data.get_question(lang).replace_placeholders(&[
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[
         ("expression", expression.to_string()),
         ("unknown_a", first_unknown.to_string()),
         ("unknown_b", second_unknown.to_string()),
@@ -380,8 +375,7 @@ fn evaluate_advanced(id: i32, lang: Language) -> Result<Problem> {
     let value_x = num_gen::integer().range(-2, -1).random();
     let value_y = num_gen::integer().range(-2, -1).random();
 
-    let problem_data = get_problem_data(id)?;
-    let question = problem_data.get_question(lang).replace_placeholders(&[
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[
         ("expression", expression.to_string()),
         ("unknown_a", first_unknown.to_string()),
         ("unknown_b", second_unknown.to_string()),