+++ backend/problems/src/ma1/percentages/identify_change_factors.rs
@@ -12,9 +12,7 @@ fn integer_increase_to_factor(id: i32, lang: Language) -> Result<Problem> {
     let increase = increase_range.random();
     let change_factor = math::utils::to_change_factor(increase);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "$100% + {increase}% = {total}% = {change_factor}$",
@@ -39,9 +37,7 @@ fn integer_decrease_to_factor(id: i32, lang: Language) -> Result<Problem> {
     let decrease = decrease_range.random();
     let change_factor = math::utils::to_change_factor(-decrease);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "$100% - {decrease}% = {total}% = {change_factor}$",
@@ -66,9 +62,7 @@ fn factor_to_increase_two_decimals(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let increase = math::utils::change_factor_to_percentage(change_factor);
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", increase)]);
     let solution = format!(
         "${change_factor} = {total}% = 100% + {increase}%$",
         total = 100 + increase
@@ -92,9 +86,7 @@ fn factor_to_decrease_two_decimals(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let decrease = math::utils::change_factor_to_percentage(change_factor).abs();
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% - {decrease}%$",
@@ -119,9 +111,7 @@ fn factor_to_large_decrease_two_decimals(id: i32, lang: Language) -> Result<Prob
     let change_factor = factor_range.random();
     let decrease = math::utils::change_factor_to_percentage(change_factor).abs();
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% - {decrease}%$",
@@ -146,9 +136,7 @@ fn factor_to_increase_one_decimal(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let increase = math::utils::change_factor_to_percentage(change_factor);
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% + {increase}%$",
@@ -173,9 +161,7 @@ fn factor_to_decrease_one_decimal(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let decrease = math::utils::change_factor_to_percentage(change_factor).abs();
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% - {decrease}%$",
@@ -200,9 +186,7 @@ fn factor_to_increase_above_2(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let increase = math::utils::change_factor_to_percentage(change_factor);
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% + {increase}%$",
@@ -227,9 +211,7 @@ fn factor_to_increase_large_number(id: i32, lang: Language) -> Result<Problem> {
     let change_factor = factor_range.random();
     let increase = math::utils::change_factor_to_percentage(change_factor);
 
-    let answer = registry::get_problem_data(id)?
-        .get_answer(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let answer = registry::get_answer(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "${change_factor} = {total}% = 100% + {increase}%$",
@@ -254,9 +236,7 @@ fn decimal_increase_to_factor(id: i32, lang: Language) -> Result<Problem> {
     let increase = increase_range.random();
     let change_factor = math::utils::to_change_factor(increase);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "$100% + {increase}% = {total}% = {change_factor}$",
@@ -281,9 +261,7 @@ fn decimal_decrease_to_factor(id: i32, lang: Language) -> Result<Problem> {
     let decrease = decrease_range.random();
     let change_factor = math::utils::to_change_factor(-decrease);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "$100% - {decrease}% = {total}% = {change_factor}$",
@@ -308,9 +286,7 @@ fn large_decimal_increase_to_factor(id: i32, lang: Language) -> Result<Problem> 
     let increase = increase_range.random();
     let change_factor = math::utils::to_change_factor(increase);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", increase)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", increase)]);
 
     let solution = format!(
         "$100% + {increase}% = {total}% = {change_factor}$",
@@ -335,9 +311,7 @@ fn large_decimal_decrease_to_factor(id: i32, lang: Language) -> Result<Problem> 
     let decrease = decrease_range.random();
     let change_factor = math::utils::to_change_factor(-decrease);
 
-    let question = registry::get_problem_data(id)?
-        .get_question(lang)
-        .replace_placeholders(&[("percent", decrease)]);
+    let question = registry::get_question(id, lang)?.replace_placeholders(&[("percent", decrease)]);
 
     let solution = format!(
         "$100% - {decrease}% = {total}% = {change_factor}$",